+++ src/actions/controlpanels/controlpanels.js
@@ -9,6 +9,7 @@ import {
   DELETE_CONTROLPANEL,
   LIST_CONTROLPANELS,
   UPDATE_CONTROLPANEL,
+  SYSTEM_INFORMATION,
 } from '@plone/volto/constants/ActionTypes';
 /**
  * Get controlpanel function.
@@ -90,6 +91,16 @@ export function updateControlpanel(url, data) {
       op: 'patch',
       path: url,
       data,
+    },
+  };
+}
+
+export function getSystemInformation() {
+  return {
+    type: SYSTEM_INFORMATION,
+    request: {
+      op: 'get',
+      path: '/@system',
     },
   };
 }