+++ src/reducers/controlpanels/controlpanels.test.js
@@ -36,6 +36,7 @@ describe('Controlpanels reducer', () => {
       },
       controlpanel: null,
       controlpanels: [],
+      systeminformation: null,
     });
   });