+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Feature
 
+- Add system information in controlpanel @iFlameing #1457
 - Added Dexterity Types Controlpanel @avoinea #1285
 
 ### Bugfix